Rename service `idpoller` to `identities-worker`
