+++ src/backend/main/go.main/facilities/REST/identities.go
@@ -119,7 +119,7 @@ func (rest *RESTfacility) CreateUserIdentity(identity *UserIdentity) CaliopenErr
 		return WrapCaliopenErr(err, DbCaliopenErr, "[CreateUserIdentity] CreateUserIdentity failed to create identity in store")
 	}
 
-	//TODO: emit nats message to IDpoller if it's a remote identity
+	//TODO: emit nats message to identities-worker if it's a remote identity
 
 	return nil
 }
@@ -183,7 +183,7 @@ func (rest *RESTfacility) PatchUserIdentity(patch []byte, userId, identityId str
 		return WrapCaliopenErrf(err4, FailDependencyCaliopenErr, "[RESTfacility] PatchUserIdentity failed with UpdateUserIdentity error : %s", err4)
 	}
 
-	//TODO: emit nats message to IDpoller if it's a remote identity
+	//TODO: emit nats message to identities-worker if it's a remote identity
 
 	return nil
 }
@@ -203,7 +203,7 @@ func (rest *RESTfacility) DeleteUserIdentity(userId, identityId string) Caliopen
 		return WrapCaliopenErrf(err2, DbCaliopenErr, "[RESTfacility DeleteUserIdentity failed to delete in store")
 	}
 
-	//TODO: emit nats message to IDpoller
+	//TODO: emit nats message to identities-worker
 	return nil
 }
